requirements.txt
autopep8==1.7.0
camelot-py==0.9.0
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
et-xmlfile==1.1.0
install==1.3.5
numpy==1.23.2
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.4.4
pdfminer.six==20220524
pycodestyle==2.9.1
pycparser==2.21
PyPDF2==2.10.4
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
toml==0.10.2
